S03E06 - Implementing the "refreshable" TwitchTeam
Not yet ready, failing test shows next thing to do.

Otherwise reorganized the main class code to be more
readable.

Next Step: implement looking at the team list for membership status,
then integrate with the Shouter, and then with the Bot to be able
to refresh upon Bot Command. Eventually hook this up to the stream
"go live" event to automatically refresh.