Episode 14: In app.js, fixed overzealous conversion of function to =>, which caused Vue.js to not see the function, further converted to the new new JS syntax thanks to CodingGarden.
Started the restructure to remove BotStatus and push that responsibility to
the Shouter. There may be some bugs as a result of this not being finished. Perhaps.

Next time: fix the incoming DTO (request) for the POST to change the Active state to be a
separate DTO from the outgoing DTO (response).