Episode 15: Completed restructuring to remove BotStatus as its behavior is now in Shouter.
Would like a way to have the integration tests not connect to Twitch, but without requiring a nested @TestConfiguration class in each test. Currently the regular @Bean in JitterShoutApplication are being called, but would like that suppressed.

Next time (still!): fix the incoming DTO (request/Command) for the POST to change the Active state to be a separate DTO from the outgoing DTO (response/Query).